requirements.txt
django
djangorestframework
pyteal==0.9.1
py-algorand-sdk==1.8.0
mypy==0.910
pytest
black==21.7b0